integer bug fix in degree conversion

+++ lib/gallary_tile.dart
@@ -162,9 +162,9 @@ class _ShowDetails extends StatelessWidget {
       isNegative = true;
       numeric = -numeric;
     }
-    int degree = numeric.floor();
-    int minute = ((numeric - degree) * 60).floor();
-    double seconds = (((numeric - degree) * 60) - minute) * 60;
+    int degree = numeric.toDouble().floor();
+    int minute = ((numeric - degree) * 60).toDouble().floor();
+    double seconds = (((numeric - degree).toDouble() * 60) - minute) * 60;
 
     return "$degree\xb0 $minute\' ${seconds.toStringAsFixed(1)}\" " +
         (latOrLong == 'lat'
